Agents/ProgressTrackingAgent/requirements.txt
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pandas==2.2.0
numpy==1.26.3
requests==2.31.0
schedule==1.2.1
pydantic==2.6.0
jinja2==3.1.3

# Optional: For vector store functionality
# pinecone-client==3.0.3
